locale-django/requirements.txt
asgiref==3.8.1
Django==3.2.25
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers
psycopg2==2.9.9
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
